requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.0
flask-migrate>=4.0.0
flask-wtf>=1.2.0
email-validator>=2.1.0
werkzeug>=3.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
country-list>=1.0.0
